:truck: Move Segement to DigitalNumber/Segement
