+++ 2-digital-clock/src/components/DigitalNumber/Segment/index.tsx
